requirements.txt
python-dotenv # to use .env file 
langchain-openai


